requirements.txt
# Data processing
pandas>=1.3.0
numpy>=1.20.0

# NLP & ML
transformers>=4.18.0
datasets>=2.0.0
torch>=1.10.0
scikit-learn>=1.0.0
spacy>=3.2.0

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.6.0

# API & Web
fastapi>=0.75.0
uvicorn>=0.17.0
pydantic>=1.9.0

# YouTube API
google-api-python-client>=2.40.0
google-auth>=2.6.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0

# MLOps
mlflow>=1.24.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Utilities
python-dotenv>=0.20.0
tqdm>=4.64.0
wordcloud>=1.8.0
beautifulsoup4>=4.10.0